Creating a Login Page with CAPTCHA Automation :
In your code we are adding a login page for user for the better interaction with the software also it will also useful for admin to handle the data for particular class/branch. 
Also In this Login Page we are adding a security features like CAPTCHA.  
Also we are adding a win32com.client API. There are several APIs available to convert text to speech in python. One of such APIs available in the python library commonly known as win32com library. It provides a bunch of methods to get excited about and one of them is the Dispatch method of the library. Dispatch method when passed with the argument of SAPI.SpVoice It interacts with the Microsoft Speech SDK to speak what you type in from the your code. 

#1 for installation of win32com.client module : 
       write this command in your terminal :  pip install pypiwin32

#2 For CAPTCHA module installation : 
        we are using a random module.